backend/requirements.txt
fastapi
uvicorn[standard]
python-nmap
python-owasp-zap-v2.4
selenium
webdriver-manager
reportlab
requests
